requirements.txt
fastapi
uvicorn
streamlit
supabase

langchain
langgraph
langchain-core
langchain-groq

python-dotenv
pandas
requests